requirements.txt
pypdf>=4.0
python-docx>=1.1
playwright>=1.40
